+++ src/opengeodeweb_viewer/vtkw_server.py
@@ -15,6 +15,7 @@ from .vtk_protocol import VtkView
 from .rpc.viewer.viewer_protocols import VtkViewerView
 from .rpc.mesh.mesh_protocols import VtkMeshView
 from .rpc.model.model_protocols import VtkModelView
+from .rpc.generic.generic_protocols import VtkGenericView
 
 
 # =============================================================================
@@ -49,10 +50,13 @@ class _Server(vtk_wslink.ServerProtocol):
         self.setSharedObject("db", dict())
 
         # Custom API
+        mesh_protocols = VtkMeshView()
+        model_protocols = VtkModelView()
         self.registerVtkWebProtocol(VtkView())
         self.registerVtkWebProtocol(VtkViewerView())
-        self.registerVtkWebProtocol(VtkMeshView())
-        self.registerVtkWebProtocol(VtkModelView())
+        self.registerVtkWebProtocol(mesh_protocols)
+        self.registerVtkWebProtocol(model_protocols)
+        self.registerVtkWebProtocol(VtkGenericView(mesh_protocols,model_protocols))
 
         # tell the C++ web app to use no encoding.
         # ParaViewWebPublishImageDelivery must be set to decode=False to match.