+++ src/tests/test_viewer_protocols.py
@@ -7,33 +7,25 @@ from opengeodeweb_viewer.rpc.viewer.viewer_protocols import VtkViewerView
 # Local application imports
 from .test_mesh_protocols import test_register_mesh
 
-class_ = VtkViewerView()
-
-
 def test_create_visualization(server):
-    server.call(class_.prefix + class_.schemas_dict["create_visualization"]["rpc"])
+    server.call(VtkViewerView.prefix + VtkViewerView.schemas_dict["create_visualization"]["rpc"])
     assert server.compare_image(3, "viewer/create_visualization.jpeg") == True
 
 
 def test_reset_camera(server):
-    server.call(class_.prefix + class_.schemas_dict["reset_camera"]["rpc"])
+    server.call(VtkViewerView.prefix + VtkViewerView.schemas_dict["reset_camera"]["rpc"])
     assert server.compare_image(3, "viewer/reset_camera.jpeg") == True
 
 
 def test_set_viewer_background_color(server):
-    server.call(
-        class_.prefix + class_.schemas_dict["set_background_color"]["rpc"],
-        [{"red": 0, "green": 0, "blue": 255}],
-    )
+    server.call(VtkViewerView.prefix + VtkViewerView.schemas_dict["set_background_color"]["rpc"], [{"red": 0, "green": 0, "blue": 255}])
     assert server.compare_image(3, "viewer/set_background_color.jpeg") == True
 
 
 def test_get_point_position(server):
     test_register_mesh(server)
-    server.call(
-        class_.prefix + class_.schemas_dict["get_point_position"]["rpc"],
-        [{"x": 0, "y": 0}],
-    )
+
+    server.call(VtkViewerView.prefix + VtkViewerView.schemas_dict["get_point_position"]["rpc"], [{"x": 0, "y": 0}])
     response = server.get_response()
     assert "x" in response["result"]
     assert "y" in response["result"]
@@ -52,14 +44,8 @@ def test_take_screenshot(server):
 
     # Take a screenshot with background jpg
     server.call(
-        class_.prefix + class_.schemas_dict["take_screenshot"]["rpc"],
-        [
-            {
-                "filename": "take_screenshot_with_background",
-                "output_extension": "jpg",
-                "include_background": True,
-            }
-        ],
+        VtkViewerView.prefix + VtkViewerView.schemas_dict["take_screenshot"]["rpc"],
+        [{"filename": "take_screenshot_with_background", "output_extension": "jpg", "include_background": True}],
     )
 
     response = server.get_response()
@@ -78,14 +64,8 @@ def test_take_screenshot(server):
 
     # Take a screenshot without background png
     server.call(
-        class_.prefix + class_.schemas_dict["take_screenshot"]["rpc"],
-        [
-            {
-                "filename": "take_screenshot_without_background",
-                "output_extension": "png",
-                "include_background": True,
-            }
-        ],
+        VtkViewerView.prefix + VtkViewerView.schemas_dict["take_screenshot"]["rpc"],
+        [{"filename": "take_screenshot_without_background", "output_extension": "png", "include_background": True}],
     )
 
     response = server.get_response()
@@ -106,14 +86,8 @@ def test_take_screenshot(server):
 
     # Take a screenshot with background png
     server.call(
-        class_.prefix + class_.schemas_dict["take_screenshot"]["rpc"],
-        [
-            {
-                "filename": "take_screenshot_with_background",
-                "output_extension": "png",
-                "include_background": True,
-            }
-        ],
+        VtkViewerView.prefix + VtkViewerView.schemas_dict["take_screenshot"]["rpc"],
+        [{"filename": "take_screenshot_with_background", "output_extension": "png", "include_background": True}],
     )
 
     response = server.get_response()